chore: move unit test to own file
